Add my solution to margin and padding exercise 2.
My initial approach was I used the borders with color white.
Since borders do not overlap, I explicitly set the bottom border to be 0px.
But my declaration was repeating.
The actual solution used the background of the card as the separator.
With this approach, padding on the card and a margin for each
section is the only thing needed.